test: move the test to the right file
