+++ test/StringField_test.js
@@ -350,6 +350,86 @@ describe("StringField", () => {
 
       expect(node.querySelector("#custom")).to.exist;
     });
+
+    it("should render a select element and it's first option is 'false', if set the enum and the default value is false.", () => {
+      const schema = {
+        type: "object",
+        properties: {
+          foo: {
+            type: "string",
+            enum: [false, true],
+            default: false,
+          },
+        },
+      };
+
+      const { node } = createFormComponent({
+        schema,
+      });
+
+      const options = node.querySelectorAll("option");
+      expect(options[0].innerHTML).eql("false");
+    });
+
+    it("should render a select element and the option's length is equal the enum's length, if set the enum.", () => {
+      const schema = {
+        type: "object",
+        properties: {
+          foo: {
+            type: "string",
+            enum: [false, true],
+            default: false,
+          },
+        },
+      };
+
+      const { node } = createFormComponent({
+        schema,
+      });
+
+      const options = node.querySelectorAll("option");
+      expect(options.length).eql(2);
+    });
+
+    it("should render a select element and the option's length is equal the enum's length, if set the enum and the default value is empty.", () => {
+      const schema = {
+        type: "object",
+        properties: {
+          foo: {
+            type: "string",
+            enum: ["", "1"],
+            default: "",
+          },
+        },
+      };
+
+      const { node } = createFormComponent({
+        schema,
+      });
+
+      const options = node.querySelectorAll("option");
+      expect(options.length).eql(2);
+    });
+
+    it("shouldn't render two empty options, when the default value is empty.", () => {
+      const schema = {
+        type: "object",
+        properties: {
+          foo: {
+            type: "string",
+            enum: [""],
+            default: "",
+          },
+        },
+      };
+
+      const { node } = createFormComponent({
+        schema,
+      });
+
+      const options = node.querySelectorAll("option");
+      expect(options.length).eql(1);
+    });
   });
 
   describe("TextareaWidget", () => {
